move tests into new credential-specific modules
