+++ sdk/identity/azure-identity/tests/test_client_secret_credential_async.py
@@ -2,16 +2,16 @@
 # Copyright (c) Microsoft Corporation.
 # Licensed under the MIT License.
 # ------------------------------------
+import asyncio
+import time
+from unittest.mock import Mock
+
+from azure.core.credentials import AccessToken
 from azure.core.pipeline.policies import ContentDecodePolicy, SansIOHTTPPolicy
 from azure.identity.aio import ClientSecretCredential
-from helpers import build_aad_response, mock_response
+from helpers import async_validating_transport, build_aad_response, mock_response, Request
 
 import pytest
-
-try:
-    from unittest.mock import Mock
-except ImportError:  # python < 3.3
-    from mock import Mock  # type: ignore
 
 
 @pytest.mark.asyncio
@@ -28,3 +28,68 @@ async def test_policies_configurable():
     await credential.get_token("scope")
 
     assert policy.on_request.called
+
+
+@pytest.mark.asyncio
+async def test_client_secret_credential():
+    client_id = "fake-client-id"
+    secret = "fake-client-secret"
+    tenant_id = "fake-tenant-id"
+    access_token = "***"
+
+    transport = async_validating_transport(
+        requests=[Request(url_substring=tenant_id, required_data={"client_id": client_id, "client_secret": secret})],
+        responses=[
+            mock_response(
+                json_payload={
+                    "token_type": "Bearer",
+                    "expires_in": 42,
+                    "ext_expires_in": 42,
+                    "access_token": access_token,
+                }
+            )
+        ],
+    )
+
+    token = await ClientSecretCredential(
+        tenant_id=tenant_id, client_id=client_id, client_secret=secret, transport=transport
+    ).get_token("scope")
+
+    # not validating expires_on because doing so requires monkeypatching time, and this is tested elsewhere
+    assert token.token == access_token
+
+
+@pytest.mark.asyncio
+async def test_cache():
+    expired = "this token's expired"
+    now = int(time.time())
+    expired_on = now - 3600
+    expired_token = AccessToken(expired, expired_on)
+    token_payload = {
+        "access_token": expired,
+        "expires_in": 0,
+        "ext_expires_in": 0,
+        "expires_on": expired_on,
+        "not_before": now,
+        "token_type": "Bearer",
+    }
+    mock_send = Mock(return_value=mock_response(json_payload=token_payload))
+    transport = Mock(send=asyncio.coroutine(mock_send))
+    scope = "scope"
+
+    credential = ClientSecretCredential("tenant-id", "client-id", "secret", transport=transport)
+
+    # get_token initially returns the expired token because the credential
+    # doesn't check whether tokens it receives from the service have expired
+    token = await credential.get_token(scope)
+    assert token == expired_token
+
+    access_token = "new token"
+    token_payload["access_token"] = access_token
+    token_payload["expires_on"] = now + 3600
+    valid_token = AccessToken(access_token, now + 3600)
+
+    # second call should observe the cached token has expired, and request another
+    token = await credential.get_token(scope)
+    assert token == valid_token
+    assert mock_send.call_count == 2
